SONAR-2706 Add extension point to customize workflow of reviews
Note that this extension point is still not exposed to public API. It's still located in
the package org.sonar.core.reviews.workflow of sonar-core.

+++ plugins/sonar-core-plugin/src/main/java/org/sonar/plugins/core/sensors/ReviewNotifications.java
@@ -77,7 +77,7 @@ public class ReviewNotifications implements BatchExtension {
     if (review.getAssigneeId() == null) { // not assigned
       return null;
     }
-    User user = userFinder.findById(review.getAssigneeId());
+    User user = userFinder.findById(review.getAssigneeId().intValue());
     return user != null ? user.getLogin() : null;
   }
 }
